Edited Layout component.
