--- src/components/Footer/Footer.scss
@@ -1,317 +1,0 @@
-@import '@styles/variables/all';
-@import '@styles/mixins/all';
-
-.container {
-  padding-right: 4.28rem;
-  padding-left: 4.28rem;
-
-  @media screen and (max-width: 1300px) {
-    padding-right: 4.28rem;
-    padding-left: 4.28rem;
-  }
-  @media screen and (max-width: 1024px) {
-    padding-right: 5rem;
-    padding-left: 5rem;
-  }
-  @include media-breakpoint-down(sm) {
-    padding-right: 60px;
-    padding-left: 60px;
-  }
-  @include media-breakpoint-down(xs) {
-    padding-right: 1.67rem;
-    padding-left: 1.67rem;
-  }
-}
-
-footer {
-  span {
-    color: #707a9b;
-    display: block;
-    font-family: $font-sofia-pro-soft;
-    font-size: 1rem;
-    margin-bottom: 15px;
-    font-weight: 700;
-    text-transform: uppercase;
-    letter-spacing: 0.5rem;
-    @include media-breakpoint-down(sm) {
-      margin-bottom: 20px;
-    }
-  }
-}
-footer .works {
-  float: left;
-  width: 50%;
-  @include media-breakpoint-down(sm) {
-    width: 100%;
-  }
-}
-
-footer .footer__logo {
-  display: flex;
-  align-items: center;
-  font-family: $font-sofia-pro-soft;
-  font-style: normal;
-  font-weight: 300;
-  font-size: 1.14rem;
-  line-height: 1.56em;
-  @include media-breakpoint-down(sm) {
-    order: 1;
-    width: 100%;
-  }
-}
-
-footer .footer__logo img {
-  margin-right: 1.42rem;
-}
-footer .works span {
-  font-family: $font-sofia-pro-soft;
-  @include media-breakpoint-down(sm) {
-    font-size: 1.6rem;
-    letter-spacing: 0.45rem;
-  }
-  @include media-breakpoint-down(xs) {
-    font-size: 1rem;
-  }
-}
-footer .works ul {
-  margin: 0;
-  padding: 0;
-  @include media-breakpoint-down(sm) {
-    margin: auto;
-    width: 100%;
-  }
-}
-footer .works ul li {
-  float: left;
-  list-style: none;
-  margin-right: 10px;
-  line-height: 2rem;
-  @include media-breakpoint-down(sm) {
-    margin-bottom: 4rem;
-  }
-  @include media-breakpoint-down(xs) {
-    margin-bottom: 1rem;
-  }
-}
-footer .works ul li a {
-  color: #fff;
-  display: block;
-  font-family: $font-sofia-pro-soft;
-  font-weight: 400;
-  font-size: 1.45rem;
-  line-height: 2.25rem;
-  position: relative;
-  letter-spacing: 0.075em;
-  text-decoration: none;
-  @include media-breakpoint-down(sm) {
-    font-size: 2.5rem;
-  }
-  @include media-breakpoint-down(xs) {
-    font-size: 1.5rem;
-  }
-}
-
-footer .works ul li a:after {
-  background: #3719ca;
-  content: '';
-  height: 2px;
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  width: 100%;
-  z-index: -1;
-  -webkit-transition: 0.2s;
-  -o-transition: 0.2s;
-  transition: 0.2s;
-}
-footer .works ul li a:hover:after {
-  height: 100%;
-}
-footer .works ul .more {
-  color: #79aeea;
-  font-family: $font-sofia-pro-soft;
-  font-size: 1.1rem;
-  line-height: 2rem;
-}
-footer .socials {
-  position: absolute;
-  left: 50%;
-  transform: translateX(-50%);
-  @include media-breakpoint-down(md) {
-    position: relative;
-    left: 0;
-    transform: translateX(0);
-    order: 0;
-    margin-left: 6.17rem;
-    // margin-bottom: 20px;
-  }
-
-  @include media-breakpoint-down(sm) {
-    margin-left: 0;
-    margin-bottom: 2.9rem;
-  }
-  // @include media-breakpoint-down(sm){
-  //     display: flex;
-  //     justify-content: center;
-  //     flex-direction: column;
-  //     margin-top: 5.5rem;
-  //     width: 100%;
-  // }
-  // @include media-breakpoint-down(xs){
-  //     margin-top: 3rem;
-  // }
-}
-footer .socials span {
-  letter-spacing: 0.14em;
-  font-size: 0.8rem;
-  @include media-breakpoint-down(sm) {
-    letter-spacing: 0.3rem;
-    font-size: 1.9rem;
-  }
-  @include media-breakpoint-down(xs) {
-    font-size: 0.9rem;
-  }
-}
-
-footer .socials ul {
-  margin: 0;
-  padding: 0;
-}
-footer .socials ul li {
-  float: left;
-  list-style: none;
-  line-height: 2rem;
-  @include media-breakpoint-down(sm) {
-    // width: 18%;
-    max-width: 6.5rem;
-  }
-  @include media-breakpoint-down(xs) {
-    max-width: 44px;
-  }
-}
-footer .socials ul li + li {
-  margin-left: 1.2rem;
-}
-footer .socials ul li a {
-  border: 1px solid #1e213d;
-  border-radius: 50%;
-  color: #fff;
-  display: block;
-  text-decoration: none;
-  width: 3.755rem;
-  height: 3.755rem;
-  font-size: 0;
-  -webkit-transition: 0.2s;
-  -o-transition: 0.2s;
-  transition: 0.2s;
-  background-position: center center;
-  background-repeat: no-repeat;
-
-  @include media-breakpoint-up(xl) {
-    width: 3.21rem;
-    height: 3.21rem;
-    background-size: 1.5rem 1.5rem;
-  }
-  @include media-breakpoint-down(md) {
-    margin: 0;
-  }
-}
-footer .socials ul li a:hover {
-  background-color: #3719ca;
-  border-color: #3719ca;
-}
-footer .socials ul li .s-beh {
-  background-image: url('./images/behance_logo.svg');
-}
-footer .socials ul li .s-dbl {
-  background-image: url('./images/dribbble_logo.svg');
-}
-footer .socials ul li .s-md {
-  background-image: url('./images/medium_logo.svg');
-}
-footer .socials ul li .s-fb {
-  background-image: url('./images/facebook_logo.svg');
-}
-footer .socials ul li .s-inst {
-  background-image: url('./images/instagram_logo.svg');
-}
-footer .socials ul li .s-twitter {
-  background-image: url('./images/twitter.svg');
-}
-
-.footer_conteiner {
-  position: relative;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  @include media-breakpoint-down(md) {
-    justify-content: flex-start;
-  }
-  @include media-breakpoint-down(sm) {
-    justify-content: center;
-    flex-direction: column;
-    // padding-top: 3.5rem;
-  }
-}
-footer {
-  padding-bottom: 1.85rem;
-  padding-top: 4rem;
-
-  @include media-breakpoint-down(sm) {
-    padding-bottom: 2.35rem;
-  }
-  @include media-breakpoint-down(xs) {
-    padding-bottom: 1.7rem;
-    padding-top: 3rem;
-  }
-  .copyright {
-    @include media-breakpoint-down(sm) {
-      justify-content: center;
-      font-size: 1rem;
-    }
-    span {
-      display: none;
-      @include media-breakpoint-down(sm) {
-        display: inline-block;
-        margin: 0;
-        font-size: 1rem;
-        color: #fff;
-      }
-    }
-    img {
-      @include media-breakpoint-down(sm) {
-        display: none;
-      }
-    }
-  }
-}
-.works {
-  margin-bottom: 0rem;
-  @include media-breakpoint-down(sm) {
-    // order: 1;
-    margin-bottom: 2rem;
-  }
-}
-.footer_container {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  @include media-breakpoint-down(sm) {
-    flex-direction: column;
-  }
-}
-
-.fb_dialog {
-  bottom: 1.85rem !important;
-  right: 2.14rem !important;
-
-  @include media-breakpoint-down(sm) {
-    right: 60px !important;
-    bottom: 2.35rem !important;
-  }
-
-  @include media-breakpoint-down(xs) {
-    right: 20px !important;
-    bottom: 1.7rem !important;
-  }
-}
